Execute Operation Criteria
Previously, the selection criteria for query versus update was on
whether the command had `SELECT` within it.  Upon encountering issues
with `SHOW TABLES`, this was changed to selection based on the existence
of `INSERT`.  This turns out to be incorrect as well since `CREATE
TABLE` doesn't work.  As part of getting a better heuristic, this change
reverts the selection criteria back to `SELECT` and allows work to start
there.

[#9]