Remove database connection configuration
Previously, the connection configuration required a user to define both a URL
and a database for connecting.  It turns out that this is both confusing and
ill-defined in the H2 documentation.  To simplify things, this change removes
the distinction requiring a user to define a complete URL, including both host
and database.  To make things a bit easier on users, this change also adds
convenience methods to the builder for in-memory and file based URLs. Finally,
the documentation is significantly updated to make the configuration
requirements more clear.

[resolves #15][#16]

Signed-off-by: Ben Hale <bhale@pivotal.io>