Test all transaction isolation levels
The default lock mode for H2 is Read Committed. A function was coded to handle
all the levels, but never adequately tested due to discussions on various
transaction support in the SPI. Transaction mutation was pulled out, but
transaction isolation kept. Hence, the other levels were never tested. This
finished out by testing them as well.

[resolves #17]

Signed-off-by: Ben Hale <bhale@pivotal.io>